backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
flask-smorest==0.41.0
Flask-SQLAlchemy==3.0.3
numpy==1.24.2
psycopg2-binary==2.9.6
PyPDF2==2.12.1
python-decouple==3.8
selenium==4.8.3
soundfile==0.12.1
SpeechRecognition==3.10.0
PyJWT==2.7.0
cryptography==41.0.1
python-dateutil==2.8.2
waitress==2.1.2